requirements.txt
requests
tqdm
Pillow
Flask
flask-cors